1. 将时间系统替换为Long；2. 修复了重新进入界面可能导致异常显示的bug

+++ easydanmaku/src/main/java/top/littlefogcat/easydanmaku/danmakus/DanmakuItem.kt
@@ -19,7 +19,7 @@ import top.littlefogcat.easydanmaku.Danmakus
  */
 class DanmakuItem(
     open var text: String,
-    open var time: Int,
+    open var time: Long,
     open var type: Int,
     open var color: Int,
     open var priority: Int,
